+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * New cache_limits option (Thanks @apzel!)
 * New simplification tolerance option (Thanks @apzel!)
+* Support environment variables in configuration file
 * Update Maputnik to 1.4
 * Add support for GDAL 25D geometry types
 * Add support for GDAL network adresses like GeoJSON services